rename to CryptoManager
